+++ plot/viewer/src/canvases/FoundationCanvas.tsx
@@ -17,7 +17,6 @@ export function FoundationCanvas(props: SketchCanvasProps) {
       shouldDrill={undefined}
       showFoldButton={false}
       injectAnchor={true}
-      applyAnchorRadialLayout={true}
     />
   );
 }